+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 ## ChangeLog
 
+* The SDK will now check metrics for infinity and NaN.  Metrics with invalid
+values will be rejected, and will result in an error logged.
+
 ## 0.1.0
 
 First release!